fix(reader): Fix ID3v2 tag lyrics size length and album cover
Fix ID3v2 tag lyrics size length and album cover:
* Fix lyrics size length not compatible on Windows x86_x64.
* Fix album cover not initialized in some situation.

+++ binding/binding.cpp
@@ -46,7 +46,7 @@ MeipuruID3v2Tag *MeipuruReadID3v2Tag(const char *filePath) {
             id3v2Tag->channels,
             id3v2Tag->length,
             id3v2Tag->lyrics.c_str(),
-            id3v2Tag->lyrics.size(),
+            (unsigned long) (id3v2Tag->lyrics.size()),
             id3v2Tag->albumCover.data,
             id3v2Tag->albumCover.size,
     };
